Discover Keystone Auth version to use in plugins
To allow these checks to work in environments with different
keystone auth versions enabled, we attempt to discover the version
of keystone auth to use based on the auth_url with 'v3' being the
only version we look for explicitly, and v2.0 being used as a
fallback. This is then used to select the correct keystone client
version (v3 or v2_0) to use when retrieving a current token or
retrieving an updated service catalog to discover the services being
checked.

We could have inspected additional ENV vars to make the determination
but that introduces and additional requirement which would complicate
reuse by a minor increment. This solution adds no such barrier.

Addresses: #598
(cherry picked from commit 6f195a7d776cd2f7da2e13d4cb8687c297670aab)