Use OSA default Keystone token provider (fernet)
The variables, keystone_token_provider and keystone_token_driver, in
user_variables set the Keystone token provider to 'uuid' and the driver
to 'sql'. These are the default values used in the Keystone project and
override the values set in openstack-ansible (OSA).

In openstack-ansible Liberty the default Keystone token provider is
'fernet' and the default token driver is 'memcache'.

Fernet tokens are recommended by a number of core members of the
Keystone project. One benefit they offer is that the tokens are not
stored in a database and so there is no requirement to periodically
remove stale tokens.

This patch removes the vars from user_variables.yml which means that the
default OSA values are used. Therefore by default rpc-openstack will
deploy Keystone using Fernet tokens.

Closes-Issue: https://github.com/rcbops/rpc-openstack/issues/873