Ship logs only from rsyslog container
Most of the logs shipped to logstash with filebeat were already on the
rsyslog container since we use rsyslog to centralize our log files on
one sever. Instead of then, once again, sending the logs to logstash
from each of the hosts and containers with the same information to
logstash, it makes sense to simply send all of the logs from the rsyslog
container to logstash.

Connects rcbops/u-suk-dev#271