Enable cinder-backup service
By default, cinder-backup is disabled in OSA. Recently, we fixed the
monitoring that we had in place for cinder-backups and discovered that
the cinder-backup service is not running.

This commit sets the necessary variable to ensure that the
cinder-backup service gets configured/deployed.

Note: by default, cinder will be configured to use swift as a backup
target. Further configuration would need to be done by the customer to
enable different backup targets.

Connected https://github.com/rcbops/rpc-openstack/issues/1308